refactor: 중복 모듈 제거, 통합
