+++ src/backend/domains/chatbot/repositories/implementations/OpenAIChatRepository.ts
@@ -5,7 +5,8 @@ import { ReservationTools } from '../../tools/ToolDefinitions';
 import { SQLiteReservationTool } from '../../tools/implementations/SQLiteReservationTool';
 import { AgentExecutor, createOpenAIFunctionsAgent } from "langchain/agents";
 import { ChatPromptTemplate } from "@langchain/core/prompts";
-import { ChatSessionEntity, ReservationInfo } from '../entity/ChatSessionEntity';
+import { ChatSessionEntity } from '../../entity/ChatSessionEntity';
+import { ReservationInfo } from '../../types/ReservationTypes';
 import { RunnableSequence } from "@langchain/core/runnables";
 import { formatToOpenAIFunctionMessages } from "langchain/agents/format_scratchpad";
 import { AgentStep } from "@langchain/core/agents";
