--- src/backend/domains/chatbot/validators/ReservationValidator.ts
@@ -1,1 +1,0 @@
-// src/backend/domains/chatbot/tools/validators/ReservationValidator.ts와 중복되므로 제거
